pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "mcp-financial-cents"
version = "0.1.0"
description = "Financial Cents REST API client and MCP server (stdio)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "msal>=1.28.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0,<2",
    "pypdf>=4.2.0",
]

[project.scripts]
financial-cents-mcp = "mcp_fc.__main__:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["financial_cents", "mcp_fc", "sharepoint"]

[tool.hatch.build.targets.sdist]
include = [
    "/financial_cents",
    "/mcp_fc",
    "/sharepoint",
    "/tests",
    "/pyproject.toml",
    "/README.md",
]
